Add method record_DLdata. Modify save_checkpoint.
Add method record_DLdata to randomly load zernike aberrations onto SLM
and record the coefficient vector and ROI. This is a function for
generating a dataset for the purposes of training a predictive model to
correct for aberrations.

Add parameter roi_only to method save_checkpoint in order to be able to
save only the roi data and none of the other metric data.